New Class Tree
Tree is the root object. ParseScriptblock still return a list of nodes.
Trees as only 1 property wich is Nodes a List<Nodes>.
Not relying on LinkedList
I thing it's for the best.

Node is know a concrete class, not abstract.

New Method: FindIndex() will return the index of the node from it's parent, either if its a root node (depth 0, no parent, parentroot is the tree)